+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.2.0]
 ### Changed
+- Only import optional dependencies (s3fs, boto3) when needed
+
+### Removed
+- Removed BlackfynnPath, which has gone unused
 
 ## [0.1.5]
 ### Added